Full capturing supports: add comments, capturing reset and atomic group.
A comment adopts the following syntax: (?#…). A capturing reset adopts the
following one: (?|…). And an atomic group the following one: (?>…).

In the same time, we update the capturing_name and comment tokens to support
escaped end-character (respectively the “>” and the “)” characters). For
example: (?<foo\>bar>…) and (?#Hello Mr (Gordon\) Freeman) are now supported.

--HG--
branch : Trunk